feat(match): Fase 3 do motor de batalha — a economia do turno

+++ internal/domain/match/resource_bar.go
@@ -12,8 +12,17 @@ package match
 // Phase 1 defines only the shape. The arithmetic — average, round price, ceiling,
 // carry-over — is Phase 3.
 type ResourceBar struct {
-	Balance int
-	Speeds  []int
+	// Balance is the standing credit or debt, and it is a float64 on purpose: it is DERIVED
+	// from an average that rarely divides evenly, and it crosses rounds, so truncating it
+	// would compound an error the rules never asked for.
+	Balance float64
+	// Speeds is the ordered list of the speeds that ACTUALLY ACTED on this bar during the current
+	// round — a speed is appended when the master opens that action, never when it is enqueued. An
+	// action still waiting in the queue is not in here, which is exactly what lets one that never
+	// reaches the price roll over to the next round with its full roll and nothing to unwind.
+	//
+	// A combined action appends to BOTH bars, because it charges both.
+	Speeds []int
 }
 
 // RecordSpeed appends a speed rolled on this bar during the current round.
